refactor(app): reuse account row pairing
