+++ packages/app/src/account/component/crypto-currency-group-accounts/crypto-currency-group-accounts.tsx
@@ -1,28 +1,15 @@
 import { View } from 'react-native';
 
-import { AccountRowInterface } from '../../interface/account-row.interface';
 import { CryptoCurrencyGroupInterface } from '../../interface/crypto-currency-group.interface';
+import { pairAccountsIntoRows } from '../../utils/pair-accounts-into-rows.util';
 import { CryptoCurrencyGroupAccountRow } from '../crypto-currency-group-account-row/crypto-currency-group-account-row';
 
 interface Props {
     readonly group: CryptoCurrencyGroupInterface;
 }
 
-const pairAccountsIntoRows = (group: CryptoCurrencyGroupInterface): AccountRowInterface[] => {
-    const rows: AccountRowInterface[] = [];
-
-    for (let index = 0; index < group.accounts.length; index += 2) {
-        rows.push({
-            left: group.accounts[index],
-            right: group.accounts[index + 1]
-        });
-    }
-
-    return rows;
-};
-
 export const CryptoCurrencyGroupAccounts = ({ group }: Props) => {
-    const rows = pairAccountsIntoRows(group);
+    const rows = pairAccountsIntoRows(group.accounts);
 
     return (
         <View className="gap-y-3">
